// src/types/instagram.d.ts
export interface VideoVersion {
  type: number;
  width: number;
  height: number;
  url: string;
  id: string;
}